fix: Cloud Brain scope naming, AST consolidation, write-path cleanup, cognitive re-queuing (#826)
* mythrax: auto-promote wisdom rule

* fix: Cloud Brain scope naming, AST consolidation, write-path cleanup, cognitive re-queuing

- Fix Cloud Brain prompt to use bare scope names ('general', 'mythrax') not CorpusName format
- Add session_id to pre_invocation boot hint in offline Brain warning
- Update REINITIALIZATION.md Phase 3 with explicit scopes and CorpusName warning
- Consolidate AST vault output to project-scoped reference/<scope>/ast/{file}_ast.md
- Remove forced header injection and post-hoc stripping in save_wiki_node_db
- Implement durable re-queuing for timed-out cognitive tasks
- Differentiate 'no extractable facts' from 'LLM failure' via no_facts_reason logging
- Fix derive_slug to use 60-char word-boundary trimming (was slicing mid-word at 50 chars)
- Build AST-derived fact DTOs from symbols when LLM unavailable
- Add sanitize_body_title_repetitions to strip H1-H4 title echoes from vault bodies
- Add E2E write-path and fact audit test suite

* fix: wire graduate MCP action, convert hypothesize/refine to async cognitive callback enqueuers

Root causes:
1. 'graduate' action was not wired in manage_handlers.rs → 500 on every Cloud Brain cycle
2. hypothesize/refine called retry_llm_json which blocked on cognitive callback poll for
   up to 300s — self-deadlock when invoked by the Cloud Brain itself

Fixes:
- Wire 'graduate' action → crate::db::graduation_pipeline::run_graduation_pipeline (pure
  cosine math, no LLM needed)
- Rewrite 'hypothesize' handler: gathers facts, builds cluster prompts in Rust, enqueues
  one Synthesis CognitiveTask per cluster (fire-and-forget, returns immediately)
- Rewrite 'refine' handler: gathers pending IdeaNodes, enqueues one Refinement
  CognitiveTask per idea/fact pair (fire-and-forget, returns immediately)
- Fix write_handlers.rs callback persistence: extract real scope from [scope:...] prompt
  prefix instead of using session_id; trigger WikiNode synthesis on Synthesis/Refinement
  task_type in addition to prompt keyword matching

* fix: resolve hypothesize clustering and graduate deserialization 500 error

- Update cluster_facts in pipeline.rs to use hybrid Jaccard token overlap similarity when dense vector embeddings are zeroed/missing
- Fix WikiNode.metacognitive_confidence type from Option<i32> to Option<f64> across contracts, pipeline, watcher, ingestion, and backend
- Update graduation_pipeline.rs SQL to cast record ID as string (SELECT *, type::string(id) AS id FROM wiki_node)
- Verify hypothesize queues 6 synthesis tasks from 312 facts and graduate completes cleanly

* fix: promote user directions to direction WikiNodes during precompact and inject into PreInvocation context

- In precompact.rs, detect directional intent in user turns ('always', 'never', 'make sure', 'don't', 'use launchctl') and save as WikiNode with node_type = 'direction' (directions/<slug>_direction.md)
- In manage_handlers.rs (handle_pre_invocation_hook), query active direction WikiNodes and render them under '### 🎯 Active User Directions' in LLM context turns
- Add explicit MEMORY SEARCH FIRST MANDATE to base_playbook preamble

+++ mythrax-core/src/contracts.rs
@@ -731,7 +731,7 @@ pub struct WikiNode {
     #[serde(default, skip_serializing_if = "Option::is_none")]
     pub temporal_range_end: Option<chrono::DateTime<chrono::Utc>>,
     #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub metacognitive_confidence: Option<i32>,
+    pub metacognitive_confidence: Option<f64>,
     #[serde(default, skip_serializing_if = "Option::is_none")]
     pub node_type: Option<String>,
     #[serde(default, skip_serializing_if = "Option::is_none")]
